+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Users module handles clients joining, renewing and leaving the queue
 * database to keep track of all current users identified by browser token
+* merged App and Api modules
 
 ## 0.4.0